gaming
main.gd: replayviewer filepath fixed

replayviewer.gd and .tscn: general updates

megamain.gd: replay instance variable, records replays if project settings say so

replay.gd: has scene_path instance variable (serializes it when saved), added functions to check if a snapshot is a full snapshot based on its relative signature, and to check if a specific snapshot from a replay is a snapshot. also added some helper funcs for accessing the "status" of a replay. (How 'long' a replay is, how far through the replay a given frame index is).

made the replay an actual plugin